+++ src/TokenSystem/TokenEventArgs/TokensBurnedEventArgs.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Numerics;
 using ContractsCore;
 using TokenSystem.Tokens;
 
@@ -6,14 +7,14 @@ namespace TokenSystem.TokenEventArgs
 {
 	public class TokensBurnedEventArgs<TTagType> : EventArgs
 	{
-		public TokensBurnedEventArgs(decimal amount, IReadOnlyTaggedTokens<TTagType> tokens, Address from)
+		public TokensBurnedEventArgs(BigInteger amount, IReadOnlyTaggedTokens<TTagType> tokens, Address from)
 		{
 			this.Amount = amount;
 			this.Tokens = tokens;
 			this.From = from;
 		}
 
-		public decimal Amount { get; }
+		public BigInteger Amount { get; }
 
 		public IReadOnlyTaggedTokens<TTagType> Tokens { get; }
