+++ src/TokenSystem/TokenManager/ITokenTagger.cs
@@ -1,3 +1,4 @@
+using System.Numerics;
 using ContractsCore;
 using TokenSystem.Tokens;
 
@@ -5,6 +6,6 @@ namespace TokenSystem.TokenManager
 {
 	public interface ITokenTagger<TTagType>
 	{
-		IReadOnlyTaggedTokens<TTagType> Tag(Address owner, decimal amount);
+		IReadOnlyTaggedTokens<TTagType> Tag(Address owner, BigInteger amount);
 	}
 }
